[PP-2840 4/5] Refactor QueryResultForwarder to support retries from both consumers and producers.
Summary:
This refactors the QueryResultForwarder to use a N-producer 1 consumer model with a watchdog per query. The watchdog keeps track of consumers and producers, and will kill the query if it doesn't see activity from either a consumer or a producer. If a new consumer calls StreamResults it cancels the previous StreamResults consumer. This allows consumers to resume a query by just calling StreamResults again.

This diff also adds a QueryID property to ExecuteScriptRequest which allows resuming of a previously disconnected query.

Test Plan: Added tests to query_result_forwarder and the query_broker server. Tested on skaffold.

Reviewers: vihang, nserrino, zasgar

Reviewed By: vihang, nserrino

JIRA Issues: PP-2840

Differential Revision: https://phab.corp.pixielabs.ai/D9310

GitOrigin-RevId: f887c523f29ace1c4e806105015730a8757a5b68