[PP-2840 2/5] Renaming QueryResultForwarder methods for clarity.
Summary:
Certain methods in the QueryResultForwarder used the term "client" which was confusing because there's really 2 clients, the Kelvins sending data to the TransferResultChunk server and the cli/ui/api receiving that data from the query broker.
This diff changes their names, so that a future diff is a bit cleaner and doesn't have to worry about the renames.

Test Plan: N/A

Reviewers: nserrino, michelle, vihang

Reviewed By: vihang

Subscribers: vihang

JIRA Issues: PP-2840

Differential Revision: https://phab.corp.pixielabs.ai/D9183

GitOrigin-RevId: c558e9d7438b1dd0a0b12da56f55e3cf1884d51c