[PP-2866] Don't hide editor text behind the minimap
Summary:
Monaco's scrollbar lets text flow behind it (as it is translucent). The Minimap has a bug where it fails to account for this. VS Code doesn't have this bug. Allowing the horizontal axis to overscroll by three columns, matching the width of the underlay space, works around it.
NOTE: Another valid workaround would be to disable the minimap entirely. These options are not mutually exclusive; we can do both if we want.

Test Plan: Open the editor for, say, `px/cluster`. Resize the editor so that it has to scroll horizontally to show everything. Notice it no longer clips behind the minimap when this happens.

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PP-2866

Differential Revision: https://phab.corp.pixielabs.ai/D9228

GitOrigin-RevId: 4db7a4c86a2e26745ee850a88509cffbc1b494bf