PP-2625: Add page to Admin view which allows user to opt in/out of UI analytics
Summary: Depends on D9250. Adds a UI page that allows users to opt in/out of UI analytics. They will need to reload (or exit out of the tab, create a new tab, etc) in order for this to take effect.

Test Plan: ran it and saw the changes reflected in DB

Reviewers: michelle, vihang, zasgar, nlanam

Reviewed By: michelle, nlanam

JIRA Issues: PP-2625

Differential Revision: https://phab.corp.pixielabs.ai/D9251

GitOrigin-RevId: 5fc14ff524d26d808310b5d5ee75194e7ac6b082