[PP-2881/PP-2880] Fix another segfault in grpc_router
Summary: Fixes a segfault that was caused by a dangling std::function being called after its bound object had been deleted. This was introduced by my changes because I got rid of an unrelated lock that happened to be held whilst the function was called, and whilst deleting the handle to the function.

Test Plan: Added test that fails on main in tsan but passes with this fix.

Reviewers: nserrino

Reviewed By: nserrino

JIRA Issues: PP-2880, PP-2881

Differential Revision: https://phab.corp.pixielabs.ai/D9299

GitOrigin-RevId: d4b9fa3cac00f17a1ce4439c21223fdb009cc402