Remove idmanager from profile service
Summary:
After moving InviteUser to authpb, we still had these leftover methods sitting around. Deprecating the profile function and removing references to idmanager as part of cleanup.

Reviewers-> Let me know if this is a fine way to deprecate a function or if I should go ahead and remove the method entirely. It's unused internally at the moment.

Test Plan: deploy public cloud with skaffold and invites still work.

Reviewers: vihang, michelle

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D9316

GitOrigin-RevId: 4892a983daf6302101a08a789c4922d2e4a2dfc3