Refactor metadata_reader_tests for better reuse
Summary: @vjain left a comment that I forgot to properly address in the diff that added this test. I ran into a bug in a later diff that would've been fixed if this was easier to maintain and the tests were combined. Therefore, decided to dedupe the tests and combine them together.

Test Plan: Ran the test, should pass

Reviewers: vihang, michelle, vjain

Reviewed By: vjain

Subscribers: vjain

Differential Revision: https://phab.corp.pixielabs.ai/D9329

GitOrigin-RevId: ab5e511192d79173eecd1ef7f248e7655a640bcb