[PC-1091] Refresh cached information when first loading it.
Summary:
*TL;DR*: Rebalances how aggressively we update cached GQL data.

Apollo has a few `fetchPolicy` options when asking it to make queries. These ones are relevant for this diff:
- `cache-first`: Query cache first. If it hits, skip the network.
- `cache-and-network`: Query cache and network simultaneously, returning cached results while waiting for network results. Updates the cache when the network call completes.
- `network-only`: Only queries the network, but still updates cache when the network call completes.

When using the `pollInterval` option on a query, the default behavior is like `cache-first` for the initial query and `cache-and-network` for polled requests.

Note: for queries like `selectedClusterInfo`, changing the query variables immediately queries the network as it should.

In the admin page, we instead want to always query the network. In some other Live View components, relying on the cache alone is sufficient. This diff rebalances the settings to match.

Test Plan: Watch graphQL calls in the dev tools when checking each admin tab, and when messing with the cluster selection in the breadcrumbs of the live view. Data should no longer start stale in most places.

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PC-1091

Differential Revision: https://phab.corp.pixielabs.ai/D9325

GitOrigin-RevId: c2b22df858053745b3fdbbc6de67084af2c11235