PC-1092 API Key claims should not fail if API Key user doesnt exist
Summary:
currently, when we create the claims for a user using an API key, we create the UserClaims for the user who created the API key. this can lead to problems if the user has been deleted. the API key should still be valid, since it is actually tied to the org.
allowed userClaims to be set as an "API user claim" using a boolean. however, we keep the user ID in the claims

Test Plan: unit tests, deployed to staging and tested API key

Reviewers: zasgar, vihang

Reviewed By: zasgar

JIRA Issues: PC-1092

Differential Revision: https://phab.corp.pixielabs.ai/D9383

GitOrigin-RevId: 92e9ca26e3af66bb9c6f9119228e7fb48c973a99