Allow `px deploy` over an already deployed vizier
Summary:
In the current state, if you run `px deploy` onto a cluster with a vizier that's already running, nothing will happen.
this is because the CRD doesn't get updated. to resolve this, we can delete the existing CRD before reapplying it. the operator properly cleans up the previous version of vizier and starts up a new one with the correct version.
this works with a skaffolded vizier as well.

Test Plan:
ran px deploy:
- on a clean clustser
- on a cluster with a vizier already running

Reviewers: vihang, nserrino

Reviewed By: nserrino

Differential Revision: https://phab.corp.pixielabs.ai/D9404

GitOrigin-RevId: 37cbca6e94c66b4c952ccd99ff83119bf31cc178