PC-1073: Remove InviteUser rpc from the protos
Summary: As @vihang and I discussed, the rpc is internal so we should just remove it eventually as a cleanup measure. I previously deprecated it for good measure, this just deletes it. Nothing should consume profilepb.InviteUser so we should be fine.

Test Plan: Ran tests and had no problems.

Reviewers: vihang, michelle

Reviewed By: michelle

Subscribers: vihang

JIRA Issues: PC-1073

Differential Revision: https://phab.corp.pixielabs.ai/D9405

GitOrigin-RevId: 4fa71911c4a9c85794da779beb40430ade22eb73