Shard vzmgr nats handlers
Summary:
We are handling all heartbeats on the same handler. The channel is only
buffered at 1024 messages. We think this is what's cauing the slow consumer
issues in the nats client and causing messages to be dropped.
This runs a buffered channel and handler goroutine per shard (256 in all)
and should fix these issues.

Test Plan:
Ran the existing tests, saw failures due to lack of sharding appropriate
mocks/calls in the test. Fixed them, tests pass.

Reviewers: zasgar, michelle

Reviewed By: zasgar

JIRA Issues: PP-2925

Differential Revision: https://phab.corp.pixielabs.ai/D9421

GitOrigin-RevId: 46fea4b713de3eb49b10ba7494b5c43576f23acd