PP-2905: Pixie Cloud should read & store new information from Vizier heartbeat
Summary:
This diff puts the hooks in Pixie cloud to store in the DB the new fields that were added to the Vizier heartbeat.
The most complex of these is the previous vizier status. When we see a Vizier status change, we want to see the previous status move to the previous_vizier_status column.

Test Plan: added unit test

Reviewers: michelle, vihang, zasgar

Reviewed By: michelle

JIRA Issues: PP-2905

Differential Revision: https://phab.corp.pixielabs.ai/D9438

GitOrigin-RevId: 963e814f137a7216879726b52fb37d6e8c45a74c