Fix metadata role to include nodes in what it watches
Summary: Ran into a bug where the k8s metadata watcher in metadata server couldn't listen to K8s because it wasn't permitted to listen to nodes. Likely a result of a recent role change for the metadata_role, although unclear why this happens. I manually edited this value and found that it fixed my problem.

Test Plan: Tested by applying on a dev instance. Eventually saw the metadata after the application and no error message in the metadata logs

Reviewers: nserrino, michelle

Reviewed By: nserrino, michelle

Subscribers: michelle

Differential Revision: https://phab.corp.pixielabs.ai/D9440

GitOrigin-RevId: 33b4e92547f9cb7e6acd9ef4806fedca0c48842c