Fix operator-deployed vizier with dev cloud
Summary:
we were seeing issues with deploying the operator version of vizier with dev cloud. this is because cloudConn now tries to use the cloudAddr in the CRD, which is incorrect.
cloudConn needs to point directly to the vzConn service. this logic still exists in the configService, however the configmap that the configService writes to is no longer used.

Test Plan: n/a

Reviewers: nserrino, vihang

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D9450

GitOrigin-RevId: d9abc005e7769e40844c6a3cd4e341c31ffbe983