[PC-1118] Round default table column width down
Summary: The default width of a column is computed as a fraction of the whole available space. Browsers try not to render thing between two pixels; Chrome was rounding up while Firefox was rounding down. This was making horizontal scroll bars (as the cells totalled a higher width than their container) appear in Chrome only.

Test Plan: Run `px/cluster` in Chrome and in Firefox. No more horizontal scrollbars with a single pixel to scroll in Chrome.

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PC-1118

Differential Revision: https://phab.corp.pixielabs.ai/D9467

GitOrigin-RevId: 7495343f9d30ce32cd826b4a83a630a82a1f66f7