PP-2877, PP-2901 Cloudconn should get status and message from CRD, if available
Summary:
the operator will be responsible for monitoring and updating the Vizier status and message in the CRD. this will allow us to show more useful states and messages to the user than the Cloudconn could do alone.
So, if the CRD exists, the cloudConn should use that status/message. otherwise, it will proceed as usual.

Test Plan: skaffold

Reviewers: nserrino, vihang, zasgar

Reviewed By: vihang

JIRA Issues: PP-2877, PP-2901

Differential Revision: https://phab.corp.pixielabs.ai/D9468

GitOrigin-RevId: 71abb61340b372453fe021384f8089ecc26757df