[PC-1124] Restore mutation modal when running trace scripts
Summary: Two bugs were present. First, since the mutation modal is a child of the canvas, it wasn't showing if the canvas was suppressed. I had introduced a faulty condition for hiding the canvas. Second bug was that `ScriptContext` was using stale data in a `setResults` call, undoing previous `setResults` calls in so doing. Fixing both restores the modal.

Test Plan: Run `px/tcp_drops` to deploy a tracepoint (you can inject `pxtrace.DeleteTracepoint` to the script right above `pxtrace.UpsertTracepoint` to make the script reset the tracepoints each run as well). Between runs (you may need to edit the script with a comment or something to force it to reload), you should now see the mutation modal once more.

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PC-1124

Differential Revision: https://phab.corp.pixielabs.ai/D9478

GitOrigin-RevId: 00b2c209193f9d86663d71b8f6ba8e0ae3d19306