PP-2875 Add statusz endpoint for cloudconnector
Summary:
this diff implements the statusz endpoint for the cloudconnector. it is able to detect whether or not it has failed to connect to cloud and whether it can run basic queries.
we can add more reasons over time, after the basic pipeline is fleshed out

Test Plan: ran skaffold and curled statusz with cloudconn in different states

Reviewers: vihang, nserrino, zasgar

Reviewed By: vihang

JIRA Issues: PP-2875

Differential Revision: https://phab.corp.pixielabs.ai/D9465

GitOrigin-RevId: d7995b5039de73e7b5ce958a8098e7a35f9be32b