Add pod DNS name to ssl certs
Summary:
we will start pinging the statusz for individual pods, and we want those communications to be SSLed.
to do so, we need to add the pod DNS resolution to our vizier certs.

Test Plan: ensure that the pod DNS resolution does exist by `kubectl exec`ing into a bash pod and curling the pod URL

Reviewers: vihang

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D9485

GitOrigin-RevId: fb13a9f469f2a02d09ceccf89c7076da21773ffa