PC-1133: Fix cache policy for cluster-details to update on refresh and occasionally poll
Summary:
Previously we had default cache policy for cluster-details which meant we always check the cache and don't have a policy to update these values. We probably want to enable users to get updated data on a refresh and polling seems to actually refresh data about broken pods.

This follows the standard cache policy we use in a number of other places.

Test Plan: Tested on dev cluster. Created a pod that failed - saw that it showed up on refresh + updated polling policy. Removed that node, then saw it disappear on refresh.

Reviewers: vihang, nlanam, michelle

Reviewed By: vihang

JIRA Issues: PC-1133

Differential Revision: https://phab.corp.pixielabs.ai/D9535

GitOrigin-RevId: 0c3fc2ae6c9e66fed82c15affc28441695355208