Increase clusterID timeout in `px deploy`
Summary:
We have been noticing that the operator-based deploys take a little longer than they used to, and result in an immature "timed out waiting for cluster ID assignment error".
This is likely because:
1. It takes some time for the operator itself to be deployed
2. The operator needs some time to pick up the deployed CRD
3. the operator now reaches out to pixie cloud for generating the YAMLs that should be applied
whereas before, the CLI used to just pull the YAMLs from GCS and apply them directly.

bumped the timeout to a time the deploy should comfortably run in

Test Plan: n/a

Reviewers: vihang

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D9539

GitOrigin-RevId: ff04b9f25cba08ec314b8a1ae2476f7b380ea992