[PP-2968 6/6] Fix doc strings for UDFs/UDAs with init args.
Summary:
Didn't add a docstring for regex_match previously so I hadn't noticed, but the DocBuilder didn't work with init args. This fixes the DocBuilder to check for args' docs for init args.
The docstring that is generated doesn't mention the fact that they're init args, which I think is fine because init args are more of an implementation detail.

Test Plan: Added tests.

Reviewers: philkuz, nserrino

Reviewed By: philkuz

JIRA Issues: PP-2968

Differential Revision: https://phab.corp.pixielabs.ai/D9579

GitOrigin-RevId: e42732940e6111160d16b72e393aca53efa0ef3a