PP-2936 Add utility for patching YAMLs
Summary:
we want to allow users to specify any patch they desire to the vizier YAMLs (most notably, taints/tolerations).
this diff adds a utility for adding patches to the YAMLs, which required cleaning up some of the templator code which iterates through the YAMLs so that it could be easily re-used.

Test Plan: to test the patch worked, called the patch function in our existing template generator which is called in our release builds. created an rc and checked that the patch was correctly applied.

Reviewers: vihang, zasgar

Reviewed By: vihang

JIRA Issues: PP-2936

Differential Revision: https://phab.corp.pixielabs.ai/D9584

GitOrigin-RevId: f53b15947fae58e3c1c2427218ba839258cfced2