Operator monitoring: Move pod data to a multi-layer map
Summary:
We previously assumed the data structure would only care about pod data that has the label `name=` in the spec and that would be our way to differentiate between running and non-running pods. We've since realized that to monitor PEMs and NATs, we need to fix this assumption.

Previously, PEM data would all be constantly overwritten on any updates because all pems have the same `name=` label value (`vizier-pem`). Also our nats pod does not have a name label on it so we were not collecting any data for it.

This attempts to fix those problems.

Test Plan: Tested on skaffold and reran the tests. Created a failing cloud conn case and a failing control plane pod case.

Reviewers: vihang, michelle

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D9632

GitOrigin-RevId: f5baa650ebf61ac9e5655af63b1cab59ab9fe750