[PP-2065 pt 1] Make table type resolution available at the same point as data type resolution.
Summary:
This diff is the first in a series of diffs to switch over to the newer TableType instead of relation.
The plan for this series of diff is as follows:
- First, resolved Table/Value types at the same point in the pipeline as data type resolution. (This is what this diff does). After this, the new type information will be available in all the same places as the old relations.
- Second, remove usage of Relation from the various IR nodes. (This will happen over diffs 2-4).
- Third, move the non-relation resolution stuff from OperatorRelationRule into the IR nodes themselves. For instance, resolving keep_input_columns for MapIRs, column mappings for unions, etc.
- Next, make op->SetRelation(...), op->relation(), etc refer to the resolved type of the op instead of the resolved relation.
- Finally, remove op->SetRelation, op->relation(), etc. This final change shouldn't change anything functionally since these functions were already made to refer to the new type system.

I think it might be easier to review these diffs if you take a quick skim of all the diffs first and then come back to them individually.

For this diff, we move ResolveTypeRule up in the analyzer to where DataTypeRule currently is. To accomplish this we have to add a `PropagateTypeChangesFromNode` method that will propagate type changes from a given node, so that rules that now run after ResolveTypeRule can still update types correctly. For instance, the ConvertMetadataRule has to propagate the change in type since it creates a conversion funcir that could potentially have a new semantic type, that the MetadataIR was unaware of.

Test Plan: Existing tests pass, skaffold works.

Reviewers: philkuz, nserrino

Reviewed By: philkuz

JIRA Issues: PP-2065

Differential Revision: https://phab.corp.pixielabs.ai/D9710

GitOrigin-RevId: 74eac3d553e2f52c27158ae84c1000ab4bc0c125