SocketTracer: Better approach to filtering non-IP traffic - Part 2
Summary:
Instead of tracing open/creat and keeping all open files in a map,
we can trace sock_sendmsg and sock_recvmsg to figure out whether the traffic is going to a socket.

This is part 2 of the diff, where we remove the old approach.

Test Plan: Test case added

Reviewers: #stirling, yzhao

Reviewed By: #stirling, yzhao

Differential Revision: https://phab.corp.pixielabs.ai/D9667

GitOrigin-RevId: 0f5f1dd6412e0324ac65185ac5305faed20bc546