PP-2964: Add detection for when cluster should be degraded when a large number of PEMs are crashing
Summary:
Detect PEM pods that are crashing. Need to check the container status of the pems to see if any are in waiting + crash loop backoff or in Terminated + Reason: Error state. Then will only show a problem if greater than 25% of the PEMs are crashing.
We can tweak this number.

Test Plan: Tested on skaffold with a forced PEM failure. Added unit test for the tester.

Reviewers: vihang, michelle

Reviewed By: vihang

JIRA Issues: PP-2964

Differential Revision: https://phab.corp.pixielabs.ai/D9646

GitOrigin-RevId: ea3aa683a5c798ec13baedd5490002e700bb71b1