SocketTracer: Better approach to filtering non-IP traffic
Summary:
Instead of tracing open/creat and keeping all open files in a map,
we can trace sock_sendmsg and sock_recvmsg to figure out whether the traffic is going to a socket.

For the moment, these new probes work in conjunction with the existing probes like read/write/send/recv,
because it is easier to get the FD from those high-level functions.

Test Plan: Test case added

Reviewers: #stirling, yzhao

Reviewed By: #stirling, yzhao

Subscribers: yzhao

Differential Revision: https://phab.corp.pixielabs.ai/D9629

GitOrigin-RevId: 74c66d9b73a3c4ca7fdc01d70738c3735eb02da0