PP-2963 Vizier is in degraded state if majority of nodes have incompatible kernel version
Summary:
this diff adds the nodeWatcher, which follows the potential new channel/status aggregation scheme discussed.
the nodeWatcher runs in a separate goroutine and separately tracks its status from the monitor itself. the status only updates when there is a node update, and the update is published to the monitor through a channel.
currently the monitor has a separate goroutine which loops through the updateChannels and tracks the last status for each.

Test Plan: deployed on skaffold, ran unit test

Reviewers: vihang, philkuz

Reviewed By: vihang

JIRA Issues: PP-2963

Differential Revision: https://phab.corp.pixielabs.ai/D9648

GitOrigin-RevId: 116652a3c20bd6faf85d83352d11e9e7587374ee