Allow MUItheme in URL query params
Summary:
we previously allowed the MUI theme to be specified in a postmessage. however, that causes a brief flicker between when the page first renders and then applies the customTheme from the postmessage.
instead, we can allow users to specify the custom theme in the URL similar with how we allow them to set the "theme" from "light"/"dark".

alternatively, another solution was to add logic so that we wait until the postMessage is received until we render the page. however, the series of postmessage calls is already complex enough and becomes harder to test. so lets keep with the simple solution for now.

the largest downside is the max length of the URL, but we shouldnt be close to hitting that.

Test Plan: yarn dev

Reviewers: vihang, nlanam, zasgar

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D9677

GitOrigin-RevId: a4ef96de99ac1de3cdf69af0ba6e24e6cc70d479