Go 1.17 Tracing Prep: Convert GoTLS probes to separate entry and return probes
Summary:
For Go 1.17 tracing, we are going to have to rely on DWARF information to identify whether values are in registers or on the stack.
Part of this requires that we use the DWARF AT_location field more consistently.

To prepare for that, start converting probes to be entry/return style. DWARF info will give us the locaiton of the arguments
on the entry probe.

Test Plan: Existing tests - no functional changes expected yet.

Reviewers: #stirling, yzhao

Reviewed By: #stirling, yzhao

Subscribers: yzhao, jps

Signed-off-by: Omid Azizi <oazizi@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D9905

GitOrigin-RevId: 12a944e2f60b92487f73ea9f9a69db407c835f7d