[PC-1167] Fix height computation and focus for graphs
Summary:
Both types of graphs could, in some cases, decide they didn't like the height of their container and slowly grow out of it. The easiest way to see this is running `px/cluster`, then setting the height of the graph to 2 (normally 3) and re-running the script.

This diff changes the container CSS for both graph types to prevent this. It also fixes the focus box when clicking on the graph so that it doesn't trigger a change in height, and removes the no-longer-needed height buffer.

Test Plan: Run `px/cluster`. Edit the vis spec to change the height of the graph from 3 to 2 (also try 1, 4, etc). The chart should reliably fit its container and not change on its own. Click the graph to focus it, and click away to unfocus it. The border should now only cover the graph, not the controls below it. It should also no longer flicker when doing this. Repeat these steps with `px/dns_flow_graph` (which uses the other type of graph from `px/cluster`).

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PC-1167

Differential Revision: https://phab.corp.pixielabs.ai/D9772

GitOrigin-RevId: 594a271e029a910a485e34e0b3b2e45ef865b320