PP-3034: Add ReconcilerPhase and LastReconcilerPhaseTimestamp to the CRD to track vizier update progress
Summary:
At the moment we track update progress in the same Phase field as the Health progress. We can't easily differentiate between an unhealthy cluster that is mid-update and one that is
truly unhealthy. We add the reconciler phase to disambiguate this information - simplifying the controller logic greatly as well as the organization of our controller vs health phase.

This diff adds the changes in the types and the CRD.

Test Plan: n/a

Reviewers: vihang, michelle

Reviewed By: vihang, michelle

JIRA Issues: PP-3034

Differential Revision: https://phab.corp.pixielabs.ai/D9783

GitOrigin-RevId: de4e5ffe5b21776e099d97947fc1e899491535ab