PP-2974 Add pre-delete hook to helm chart
Summary:
there is currently a race condition in the helm chart, where the operator is deleted before it can pick up the CRD delete.
as a result, pixie pods are left running in the cluster.
to resolve this, we can use a pre-delete hook which deletes the pixie pods. (didnt use post-delete, since we need to use some of the RBAC rules that Pixie has).

Test Plan: built operator rc, deployed helm chart and deleted helm chart. confirmed that pixie pods were deleted (minus nats, because it doesnt fall into the label, but that will be fixed soon)

Reviewers: vihang, philkuz, zasgar

Reviewed By: vihang

JIRA Issues: PP-2974

Differential Revision: https://phab.corp.pixielabs.ai/D9779

GitOrigin-RevId: 33b68ecce5cb86db301c9f0d834787c9a0055823