Extra uprobe trigger example
Summary:
This is a test that shows an unexpected behavior with uprobes that can occur when two seemingly
different files, which share the same inode, are probed.

Note this is not a case of symlinks, this is actually a case where two separate files which share the same inode are probed.

This can occur if a container is run twice on the same host.

We will need a fix for this in the future.

Test Plan: This is a test that shows the faulty behavior.

Reviewers: #stirling, yzhao

Reviewed By: #stirling, yzhao

Subscribers: yzhao

Differential Revision: https://phab.corp.pixielabs.ai/D9791

GitOrigin-RevId: 22684c0a6997c4688c4606ec4f0fbc05ee183954