PP-3037: Vizier sends over Updating health status if Vizier.ReconcilerPhase is ever in the Updating phase
Summary:
If the Vizier.ReconcilerPhase is ever in the Updating Phase, we want that to supersede the VizierStatus. Given that we will set the ReconcilerStatus to failed when ttl expires this should ensure
that we dont' show an unhealthy state when the cluster is really mid-update.

Test Plan: Tested on skaffold. Forced the operator to just set updating and cloud only received updating from the cluster.

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PP-3037

Differential Revision: https://phab.corp.pixielabs.ai/D9789

GitOrigin-RevId: ada16cf2de0db5cfb146b2db4f3a9014bcf8d7b9