PP-3035: Create a watcher that checks for Viziers stuck Updating.
Summary:
Reconciler can get stuck updating, fail mid-way through, etc.
We want to make sure that if the Updating stage is stuck for a long time we can recover. This diff adds a watcher that checks every 3 minutes for Viziers that are stuck in the Updating stage for longer than a timeout period.
If they are stuck for longer, they are marked as ReconciliationPhase Failed.

Test Plan: Tested on skaffold, forced a stuck updating cluster and got the retry to occur.

Reviewers: vihang, michelle

Reviewed By: vihang, michelle

JIRA Issues: PP-3035

Differential Revision: https://phab.corp.pixielabs.ai/D9790

GitOrigin-RevId: 4e33b311a7c3616b309fe9a5c193482ac2ef0bcc