PP-3023: Metadata service should send Service updates including ClusterIP to Kelvin
Summary: We want metadata UDFs to have access to the service Cluster IP. This information is not currently being sent to Kelvin, but this diff changes that so that it will now be sent over. Next up: Have Kelvin UDFs use the ClusterIP to resolve remote_addr in the event that the traffic is traced client-side.

Test Plan: updated unit test

Reviewers: michelle, oazizi, zasgar

Reviewed By: michelle

JIRA Issues: PP-3023

Differential Revision: https://phab.corp.pixielabs.ai/D9795

GitOrigin-RevId: 0df6f3c93b6b7d93fb74f8e1d943c57412b23724