PP-3023: px.ip_to_service_id should be a Kelvin-only UDF
Summary: While doing another task in this area, I discovered that while px.ip_to_pod_id is correctly annotated as a Kelvin-only UDF, px.ip_to_service_id (which is a layer on top of the other one) is missing that annotation. It should be a Kelvin-only UDF so this diff fixes that.

Test Plan: n/a

Reviewers: michelle, oazizi, zasgar

Reviewed By: michelle

JIRA Issues: PP-3023

Differential Revision: https://phab.corp.pixielabs.ai/D9794

GitOrigin-RevId: 05965d1b5930c96b81b52694855f7ef90fcefb40