PP-3023: Add service IPs to metadata state
Summary: We want to track service IPs in addition to pod IPs when when we resolve a remote address to a particular Kubernetes service. This diff takes service updates containing service IP information and propagates it to our metadata state. The next step is to update the px.ip_to_service_id UDF in order to fully close the loop on this information. We also may want to consider adding UDFs to fetch external IP and cluster IP for a given service.

Test Plan: updated existing unit tests

Reviewers: oazizi, michelle

Reviewed By: michelle

JIRA Issues: PP-3023

Differential Revision: https://phab.corp.pixielabs.ai/D9816

GitOrigin-RevId: 6415340dc8b2284ad89cbc0dc33272783a63ca1a