PP-3045: Update the spec with default values after setting up deploy parameters
Summary: We set default values for vz during `deploy()` but don't write them to the spec. We assume that these values are set for other parts of the code and have run into some issues. This is a critical change that will probably resolve some unhealthy clusters to healthy immediately as soon as we update Vizier for users.

Test Plan: Tested on skaffold with a cluster that would get classified as not PVC compatible (and therefore automatically set useEtcdOperator set to true).

Reviewers: vihang, michelle, nserrino

Reviewed By: vihang

JIRA Issues: PP-3045

Differential Revision: https://phab.corp.pixielabs.ai/D9820

GitOrigin-RevId: 912e3566311a87855c950f8d3a1067013a52d957