PC-1179: JSON array of service should deep link in UI to first element in list
Summary: For some records, we have an array of services like ["pl/nats", "pl/stan"]. For those records, we don't know which service to attribute the measurement to. This presents a problem with some of the deep linking in the UI. When you double click an array service in the service graph, the url generator expects the non array form of service. This diff fixes that, by checking to see if the service is an array (like we do in deep links in tables), and if it is, take the user to the first of the services in the array. In the future we will want to allow the user to select which service in the array, but this is an improvement for now.

Test Plan: updated unit test

Reviewers: michelle, nlanam, zasgar

Reviewed By: nlanam

JIRA Issues: PC-1179

Differential Revision: https://phab.corp.pixielabs.ai/D9832

GitOrigin-RevId: 80b27e1a6faa4712a7833025b2cefd0e322223c7