PC-1175: px/cluster,px/namespace to list all Services, including non HTTP
Summary: Depends on D9864. Previously, we were doing an inner join between the process_stats table and http_events table to generate the list of services present. Now that PC-1174 is fixed, we can do a left join, and show non-HTTP services even when we have HTTP-specific columns. Also, HTTP-specific columns in Services list are now prepended with HTTP to make it clear that these are HTTP statistics, not overall statistics for the service.

Test Plan: ran it in UI

Reviewers: philkuz, oazizi, htroisi, zasgar

Reviewed By: oazizi

JIRA Issues: PC-1175

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D9870

GitOrigin-RevId: edbc69049b6c23d031d27ea74b978d5f13d62aff