PP-3063 Add gql endpoint for fetching IDE config path
Summary:
On right-click (or other interaction), the UI should show the user a dropdown of their configured IDEs and the path to the selected symbol.
there were two possibilities i could think of for the gql endpoint:
1. (chosen in this diff) Upon right click, make a call to our backend with the symbol. the backend returns the IDE paths.
2. Preload the orgs configured IDE settings. Upon right click, fill in the IDE paths with the symbol in the UI.
I figured that (1) provides more flexibility for configuring the IDE path/keeps unnecessary logic out of the UI. Our paths are also expected to be go templates, so it will be easier to fill out there.

Test Plan: existing tests pass/build

Reviewers: vihang, zasgar

Reviewed By: vihang

JIRA Issues: PP-3063

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D9886

GitOrigin-RevId: e7ca6cfa0bab6bf4fd842ab1bf0b2d22647fc17e