PP-3072 Fix bug where operator occassionally fails to Update()
Summary:
We've noticed an issue where the operator fails to update the Vizier CRD (most commonly seen on vizier updates):
```
ime="2021-10-05T22:03:21Z" level=info msg="Failed to update Vizier instance" error="Operation cannot be fulfilled on viziers.px.dev \"pixie\": the object has been modified; please apply your changes to the latest version and try again"
```

This can leave the vizier in a stuck updating state, even when it is healthy.
I was able to narrow it down to the part of the code where we're updating the Vizier, and we wait for the cluster to become healthy. Depending on the amount of time this takes, we have many other goroutines potentially updating the CRD in the background. Instead, we should refetch the vizier right before we're about to make an update to it.

This doesn't prevent all race conditions, however, as something could update the Vizier CRD right after we refetch it. If we see more issues here, we could potentially add a retry/backoff.

Test Plan: built operator rc, did px deploys and px updates.

Reviewers: philkuz, vihang

Reviewed By: vihang

JIRA Issues: PP-3072

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10005

GitOrigin-RevId: 4fe5316b5a9e292219a14f6a2a0c9a811cabc7c1