PP-3083: Add analyzer rule that redacts sensitive columns from being accessible.
Summary:
Sensitive columns will have information that we might want to restrict from certain users. We will be pushing for a coarse-grained redaction method this week, but I tried to design the interface to enable more fine-grained permissions in the future. We can expand the logic used in the "RestrictedColumn" function - ie RBAC-ing columns.

This diff does not:
1. incorporate the code into  the analyzer.
2. Define restricted columns.
3. Propogate whether to restrict columns to the user.

Test Plan: Added tests on the skaffold

Reviewers: jamesbartlett, nserrino, vihang

Reviewed By: jamesbartlett

JIRA Issues: PP-3083

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10015

GitOrigin-RevId: a8f87b0592ee15a27d6ebaab7f862e62a7155898