PP-3083: Pipe the DataAccess CRD for use in the compiler
Summary:
We pass the DataAccess field info from the Vizier CR over to the actual compiler. I chose to only pass the boolean value of which columns we want to keep.

An alternative and a future option is to pass a mapping of the table names to sensitive column names. We also could have passsed this value inside of the Schena object as well. Given that we're on a tighter time budget to get this out, I chose a simple and reversible option for now that opens the door for a simple change in the future.

Test Plan: Tested on skaffold and updated tests to reflect the API changes.

Reviewers: vihang, michelle, jamesbartlett

Reviewed By: michelle

JIRA Issues: PP-3083

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10023

GitOrigin-RevId: e65dd9642c8c874d230cc5fc1cf6843ba12f75ae