PP-3067 Fix bug where IPs are incorrectly expired from IP map
Summary:
We track IP -> PodInfo in a map. What we saw was that the IP that could not be resolved wasn't in this map, although strangely Kelvin knew about the pod.
We also clean up expired metadata in the PEMs/Kelvin if the pod has been terminated for over 24h.

This is the sequence of events that lead to the issue:
1. Pod A starts up with podIP. We add podIP -> podA in the IP map.
2. Pod A with podIP terminates.
3. Pod B later starts up with podIP, we update podIP -> podB in the IP map.
4. 24 hours later, Pod A is cleaned up, along with the podIP in the map because we don't check which pod is in the map!

Test Plan: added unit test

Reviewers: vihang, nserrino, zasgar

Reviewed By: vihang

JIRA Issues: PP-3067

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10020

GitOrigin-RevId: d77a01d3e935f49c5048a4362b6f153d8f7d2df5